Appending color swatches working, still refactoring
